+++ CHANGELOG.md
@@ -14,6 +14,7 @@ All notable changes to this library are documented in this file.
 
 - Update bundled `states` geographies to include territories of `AS`, `PR`,
 `VI`, `GU`, and `MP`.
+- Implement tweet message truncation for SPS messages (#393).
 
 ## **1.4.0** (9 March 2021)